+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes in v1.1b1
 - Added support for Windows 10 and higher.
 - Added `PAPPL_SOPTIONS_NO_TLS` option to disable TLS support.
 - Added Wi-Fi callbacks to support configuration over IPP-USB (Issue #45)
+- Added buttons and sub-commands to pause and resume printers (Issue #124)
 - `papplMainLoop` now uses a persistent location for state and spool files by
   default (Issue #128)
 - `papplMainLoop` now supports clients talking to a system-wide server running